refactor: introduce base message model
