+++ src/services/messaging/models/__init__.py
@@ -1,1 +1,3 @@
- 
+from .base_message import BaseMessage
+
+__all__ = ["BaseMessage"]
